refactor: GetRoot() と setupTestRepo を worktree.go から分離
worktree.go に定義されている GetRoot() は config.go と config_cmd.go
から呼ばれており、worktree パッケージ削除後も生存が必要なため repo.go
に移動。setupTestRepo ヘルパーは remote_test.go からも参照されるため
testutil_test.go に移動。GetRoot テストは repo_test.go に移動。

振る舞い変更なし（純粋リファクタリング）。

Refs: #52
